Ticket #382 Implement RFC 43 - Support Search and Selection on Double Identity Properties Submission for 2.0 stream.  1.2 stream to follow.
git-svn-id: http://svn.osgeo.org/mapguide/trunk/MgDev@2686 a442ad82-b40c-0410-ae44-e6d7bd6e4a36

+++ Server/src/Services/Feature/ServerGetProviderCapabilities.cpp
@@ -872,7 +872,9 @@ bool MgServerGetProviderCapabilities::Initialize()
     s_FdoDataType[FdoDataType_Boolean]     = "Boolean";
     s_FdoDataType[FdoDataType_Byte]        = "Byte";
     s_FdoDataType[FdoDataType_DateTime]    = "DateTime";
-    s_FdoDataType[FdoDataType_Decimal]     = "Decimal";
+    // Implementation Note:  FdoDataType_Decimal is currently mapped to MgPropertyType::Double.
+    // An MgDecimalProperty class should be implemented in a future release.
+    s_FdoDataType[FdoDataType_Decimal]     = "Double";
     s_FdoDataType[FdoDataType_Double]      = "Double";
     s_FdoDataType[FdoDataType_Int16]       = "Int16";
     s_FdoDataType[FdoDataType_Int32]       = "Int32";
